requirements.txt
python-telegram-bot>=21.0,<22.0
instaloader>=4.14
requests>=2.32
Pillow>=10.4
imageio>=2.36
imageio-ffmpeg>=0.5
numpy>=2.1
opencv-python-headless>=4.10
python-dotenv>=1.0
filelock>=3.15
browser-cookie3>=0.19
pytest>=8.3
